LPD-11253 Typescript types
