+++ modules/apps/journal/journal-web/types/src/main/resources/META-INF/resources/js/translation_manager/Types.d.ts
@@ -1,0 +1,15 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {Locale} from 'frontend-js-components-web';
+export interface TranslationManagerProps {
+	defaultLanguageId: Liferay.Language.Locale;
+	fields: Fields;
+	locales: Locale[];
+	namespace: string;
+	selectedLanguageId: Liferay.Language.Locale;
+}
+export declare type Field = Record<Liferay.Language.Locale, string>;
+export declare type Fields = Record<string, Field>;
